Closes #14 - Refactor: Nav 컴포넌트 로그인 유형에 따라 구분
