+++ components/Nav/Icon.tsx
@@ -1,17 +1,18 @@
-import styles from "../../styles/Nav/Icon.module.scss";
+import styles from "../../styles/components/Nav/Icon.module.scss";
 import Image from "next/image";
 
 interface MyProps {
 	children: string;
 	src: string;
 	alt: string;
+	clicked: string;
 }
 
 const Icon: React.FC<MyProps> = (props) => {
-	const { children, src, alt } = props;
-
+	const { children, src, alt, clicked } = props;
+	const iconBtnClassName = `${styles["icon-btn"]} ${styles[clicked]}`;
 	return (
-		<button className={styles["icon-button"]} type="button">
+		<button className={iconBtnClassName} type="button">
 			<Image
 				className={styles["icon-img"]}
 				src={src}
