--- styles/Nav/IconGroup.module.scss
@@ -1,10 +1,0 @@
-@import "../mixin";
-@import "../variables";
-
-.icon-group {
-	@include flex-row();
-	margin-right: 20rem;
-	button:first-of-type {
-		margin-right: 1.75rem;
-	}
-}
