+++ styles/components/Nav/SellerTopNavBar.module.scss
@@ -1,0 +1,21 @@
+@import "../mixin";
+@import "../variables";
+
+.nav-wrap {
+	@include flex-row();
+	@include align-center();
+	width: 100%;
+	height: 5.625rem;
+	background: #ffffff;
+	-webkit-box-shadow: 0px 4px 5px rgba(0, 0, 0, 0.1);
+	box-shadow: 0px 4px 5px rgba(0, 0, 0, 0.1);
+}
+
+.logo {
+	margin-left: 6.25rem;
+	margin-right: 1rem;
+}
+
+.title {
+	@include font(500, 30px, 38px);
+}
