+++ styles/components/Nav/TopNavBar.module.scss
@@ -1,13 +1,20 @@
 @import "../mixin";
 @import "../variables";
 
-.header-wrap {
+.nav-wrap {
 	@include flex-row();
-	@include justify-between();
+	@include justify-around();
 	@include align-center();
 	width: 100%;
 	height: 5.625rem;
 	background: #ffffff;
 	-webkit-box-shadow: 0px 4px 5px rgba(0, 0, 0, 0.1);
 	box-shadow: 0px 4px 5px rgba(0, 0, 0, 0.1);
+}
+
+.nav-icon-group {
+	@include flex-row();
+	& button:first-of-type {
+		margin-right: 1.75rem;
+	}
 }
